+++ ezdxf/lldxf/packedtags.py
@@ -210,7 +210,6 @@ class VertexArray(PackedTags):
             yield DXFVertex(self.VERTEX_CODE, point)
 
     def append(self, point):
-        # LWPolylinePoints does not maintain the point count attribute!
         if len(point) != self.VERTEX_SIZE:
             raise DXFValueError('point requires exact {} components.'.format(self.VERTEX_SIZE))
         self.value.extend(point)